refactor: Passkeys auth in a single place
