+++ app/views/gone.scala.html
@@ -1,7 +1,7 @@
 @import com.gu.googleauth.UserIdentity
 @import com.gu.janus.model.JanusData
-@import models.DisplayMode
 @import play.api.Mode
+@import views.html.helper.CSPNonce
 
 @(user: UserIdentity, janusData: JanusData)(implicit req: RequestHeader, mode: Mode, assetsFinder: AssetsFinder)
 
@@ -13,4 +13,10 @@
             <p>Please go back to the <a href="@routes.Janus.index">home page</a> and follow the new link from there.</p>
         </div>
     </div>
+
+    @if(mode == Mode.Dev) {
+        <script src="https://janus-frontend.local.dev-gutools.co.uk/passkeys-auth.js" @{CSPNonce.attr}></script>
+    } else {
+        <script src="@assetsFinder.path("frontend/passkeys-auth.js")" @{CSPNonce.attr}></script>
+    }
 }
