+++ frontend/passkeys-auth.js
@@ -1,0 +1,13 @@
+import { getCsrfTokenFromMetaTag } from "./utils/csrf.js";
+import { authenticatePasskey } from "./utils/passkeys.js";
+
+document.addEventListener("DOMContentLoaded", async () => {
+  try {
+    await authenticatePasskey(
+      window.location.pathname + window.location.search,
+      getCsrfTokenFromMetaTag(),
+    );
+  } catch (error) {
+    console.error("Failed to authenticate passkey on gone page:", error);
+  }
+});
